Refactor of the submission utility (#554)
The submission utility in the application has become a mess since
it's creation due to a limited view of its usage at conception
and extending its functionality over time.

This overhaul of the utility accomplishes the following:

- Moves from passing call back functions to using async utility functions
- Moves pipeline-related utility functions into pipeline-editor (Part of #494)
- Update all calls to metadata endpoints to use the metadata service. Fixes #271
- Move ErrorDialogContent into ui-components. The first step towards #403
- Update function, class, and file names to match actual functionality
 - Modularized HTTP_STATUS success code verification. Fixes #232

Fixes #400
