+++ elyra/pipeline/handlers.py
@@ -72,18 +72,13 @@ class PipelineSchedulerHandler(HttpErrorMixin, APIHandler):
     def post(self, *args, **kwargs):
         self.log.debug("Pipeline SchedulerHandler now executing post request")
 
-        try:
-            pipeline_definition = self.get_json_body()
-            self.log.debug("JSON payload: %s", pipeline_definition)
+        pipeline_definition = self.get_json_body()
+        self.log.debug("JSON payload: %s", pipeline_definition)
 
-            pipeline = PipelineParser.parse(pipeline_definition)
+        pipeline = PipelineParser.parse(pipeline_definition)
 
-            response = PipelineProcessorManager.process(pipeline)
-            json_msg = json.dumps(response.to_json())
-        except (ValueError) as se:
-            raise web.HTTPError(400, str(se))
-        except Exception as ex:
-            raise web.HTTPError(500, repr(ex))
+        response = PipelineProcessorManager.process(pipeline)
+        json_msg = json.dumps(response.to_json())
 
         self.set_status(200)
         self.set_header("Content-Type", 'application/json')
