+++ packages/application/package.json
@@ -30,10 +30,12 @@
     "watch": "tsc -w"
   },
   "dependencies": {
+    "@elyra/ui-components": "^0.11.0-dev",
     "@jupyterlab/apputils": "^2.0.2",
     "@jupyterlab/coreutils": "^4.0.2",
     "@jupyterlab/ui-components": "^2.0.2",
-    "@jupyterlab/services": "^5.0.2"
+    "@jupyterlab/services": "^5.0.2",
+    "react": "^16.8.6"
   },
   "devDependencies": {
     "rimraf": "~3.0.0",
