+++ packages/application/src/index.ts
@@ -18,4 +18,4 @@ import '../style/index.css';
 
 export * from './parsing';
 export * from './services';
-export * from './submission';
+export * from './requests';
