+++ packages/pipeline-editor/tsconfig.json
@@ -18,6 +18,9 @@
   "references": [
     {
       "path": "../application"
+    },
+    {
+      "path": "../ui-components"
     }
   ]
 }
