+++ packages/ui-components/src/index.ts
@@ -13,5 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+export * from './ExpandableErrorDialog';
 export * from './ExpandableComponent';
 export * from './icons';
