+++ tsconfig-base.json
@@ -5,6 +5,7 @@
     "declaration": true,
     "esModuleInterop": true,
     "incremental": true,
+    "inlineSources": true,
     "jsx": "react",
     "module": "esnext",
     "moduleResolution": "node",
